Move MerkleStream to submodule
